Use the contingent interface.
Following advice from @gselzer:

The Contingent interface is used to prevent Op Matching based on factors
other than input types, e.g. Number of Dimensions, like you had before.
If we keep the Contingent then it will simply overlook the op if we
input a non-2D image, instead of attempting the op and throwing an error.

Let's say that, in the future, we want to make a Hough Circle/Sphere
Transform that takes in only 3 Dimensions, or something else along
these lines that does not involve the op signature itself. We would
have to refactor this op instead of make a new one, which goes against
the extensible case-logic that Ops should be.

The run-time checking could potentially be time consuming. Let's say
that instead of using ops().run() the user calls ops().op(), which
would return an Op reference that the user could use to run through
a bunch of images, instead of calling ops().run() with every image.
While the user could theoretically pass an image into this reference
that would not be caught by conforms, it is part of the social contract
with the user that they will not do this. And if there are a lot of
images run over this reference, then it could hamper run time if the
reference is used many times.